% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_install.R
\name{cran_install}
\alias{cran_install}
\title{install package from CRAN with typo suggestions}
\usage{
cran_install(package, max_distance = 3, ...)
}
\arguments{
\item{package}{Name of the package to install}

\item{max_distance}{maximum distance. See \code{\link[stringdist]{stringdist}}}

\item{\dots}{Passed to \code{\link[utils]{install.packages}}}
}
\value{
nothing
}
\description{
Attempts to install a package from CRAN, and suggests alternatives
if it not available
}
\examples{
\dontrun{
  cran_install( "dlpyr" )
  cran_install( "google" )
}
}

